// lib/api/schemas/potentialType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Physense API
 * Physics simulation backend for the Physense platform.
 * OpenAPI spec version: 0.1.0
 */

export type PotentialType = typeof PotentialType[keyof typeof PotentialType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PotentialType = {
  harmonic: 'harmonic',
  infinite_well: 'infinite_well',
  finite_well: 'finite_well',
  barrier: 'barrier',
  step: 'step',
  double_well: 'double_well',
  free: 'free',
  custom: 'custom',
} as const;